requirements.txt
python-dotenv==1.0.0
folium==0.15.1
openrouteservice==2.3.3
geopy==2.4.0
streamlit_folium==0.18.0
pandas==2.1.1
requests==2.31.0
